Don't strip characters form the search word
If someone marked a selection and requests to search for it, we should
have faith in that they selected the correct thing they want to search
for and not try to be smarter by removing an arbitrary list of symbols.

For example, the ClojureScript has symbols like `clj->js` and we
previously stripped the `>`. I know that Haskell uses a plethora of
symbols for its operators and those are indexed as-is in the docset.