Clear handler after the list was shown
`run` is not executed before `input` if the command was invoked from the
command palette and not a key binding. However, because the input method
is invoked twice before the handler is actually used, we cannot clear it
on the first invocation of `input`, so we add a callback that will clear
the handler only if a list is being shown.